# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  db.WaasWallet:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      network:
        type: string
      user_id:
        type: string
      waas_wallet_id:
        type: string
    type: object
...
  handlers.CreateWalletRequest:
    properties:
      network:
        description: 'Network is the blockchain network. Supported: BTC BCH LTC BSC
          ETH POL TRX SOL XRP'
        enum:
        - BTC
        - BCH
        - LTC
        - BSC
        - ETH
        - POL
        - TRX
        - SOL
        - XRP
        example: ETH
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        properties:
          code:
            example: VALIDATION_ERROR
            type: string
          message:
            example: phone is required
            type: string
        type: object
      success:
        example: false
        type: boolean
    type: object
...
  handlers.WalletListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.WalletData'
        type: array
      success:
        example: true
        type: boolean
    type: object
  handlers.WalletResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.WalletData'
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: engineering@rach.finance
    name: Rach Finance Engineering
  description: |-
    DigitalFX is a multi-currency fintech platform for Central/West Africa.
    It combines fiat accounts (XAF, XOF, USD, GBP, EUR), HD crypto wallets (BIP-44 via Rach WaaS),
    and stablecoin P2P transfers (USDC/USDT via Rach CaaS ERC-4337 SCWs).
    Mobile Money deposits and withdrawals are handled via HUB2 (MTN, Orange).
  title: DigitalFX API
  version: "1.0"
paths:
...
  /wallets:
    get:
      description: Returns all on-chain HD wallets (BIP-44) provisioned for the authenticated
        user via the Rach WaaS service.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WalletListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List HD wallets
      tags:
      - WaaS - Crypto Wallets
    post:
      consumes:
      - application/json
      description: Provisions an HD wallet seed (if not already created) and derives
        a deposit address for the given blockchain network. Idempotent — safe to call
        again for the same network.
      parameters:
      - description: Network selection
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateWalletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.WalletResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create HD wallet for a network
      tags:
      - WaaS - Crypto Wallets
  /wallets/{walletId}/address:
    get:
      description: Returns the deposit address details for the specified wallet ID.
      parameters:
      - description: Wallet UUID
        in: path
        name: walletId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.WaasWallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get deposit address
      tags:
      - WaaS - Crypto Wallets
  /wallets/addresses:
    get:
      description: Returns every derived on-chain address for the caller across all
        networks, each with its live per-currency balances (native coin + on-chain
        stablecoins). Pass ?refresh=true to force a live on-chain balance refresh
        (slower, always accurate).
      parameters:
      - description: Force a live on-chain balance refresh
        in: query
        name: refresh
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payments.ListCustomerAddressesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List WaaS addresses & balances
      tags:
      - WaaS - Crypto Wallets
...
  /wallets/deposit:
    post:
      consumes:
      - application/json
      description: Initiates a Mobile Money collection (deposit) via HUB2, crediting
        the user's fiat account when confirmed. Returns the HUB2 reference for webhook
        correlation.
      parameters:
      - description: Deposit details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.DepositRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.Hub2RefResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Initiate Mobile Money deposit
      tags:
      - Mobile Money - HUB2
  /wallets/estimate-gas:
    post:
      consumes:
      - application/json
      description: Estimates the network fee for an EVM transfer (BSC, ETH, POL).
        Amount is in wei.
      parameters:
      - description: Gas estimate request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.EstimateGasRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payments.GasEstimate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Estimate on-chain gas (WaaS)
      tags:
      - WaaS - Crypto Wallets
  /wallets/export-key:
    post:
      consumes:
      - application/json
      description: Exports the private key for one of the caller's derived WaaS addresses.
        HIGHLY SENSITIVE — grants full control of that address. Index must match the
        derivation index used when the address was created (default 0). Gate behind
        step-up auth before surfacing in any UI.
      parameters:
      - description: Network + derivation index
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ExportKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payments.ExportPrivateKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export WaaS private key (SENSITIVE)
      tags:
      - WaaS - Crypto Wallets
...
  /wallets/seed:
    get:
      description: Returns the BIP-39 mnemonic for the caller's Rach WaaS HD wallet.
        HIGHLY SENSITIVE — anyone with this phrase controls ALL of the user's crypto.
        Intended for self-custody export; gate behind step-up auth (PIN/biometric/2FA)
        before surfacing in any UI.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payments.GetSeedPhraseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reveal WaaS seed phrase (SENSITIVE)
      tags:
      - WaaS - Crypto Wallets
...
  /wallets/swap/execute:
    post:
      consumes:
      - application/json
      description: Broadcasts a swap transaction from the caller's WaaS wallet using
        a previously quoted route.
      parameters:
      - description: Swap details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ExecuteSwapRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payments.ExecuteSwapResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Execute a WaaS token swap
      tags:
      - WaaS - Crypto Wallets
  /wallets/swap/history:
    get:
      description: Returns the caller's paginated swap transaction history.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Results per page (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payments.GetSwapHistoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get WaaS swap history
      tags:
      - WaaS - Crypto Wallets
  /wallets/swap/quote:
    get:
      description: Returns a rate quote for swapping tokens across chains using WaaS
        swap.
      parameters:
      - description: Source chain
        in: query
        name: from_chain
        required: true
        type: string
      - description: Destination chain
        in: query
        name: to_chain
        required: true
        type: string
      - description: Source token
        in: query
        name: from_token
        required: true
        type: string
      - description: Destination token
        in: query
        name: to_token
        required: true
        type: string
      - description: Amount in base units
        in: query
        name: amount_in
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payments.SwapQuoteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a WaaS token swap quote
      tags:
      - WaaS - Crypto Wallets
  /wallets/transactions:
    get:
      description: Returns the caller's on-chain transaction history from Rach WaaS,
        filterable by network, currency and status. This is real blockchain history,
        distinct from the CaaS iUSD ledger.
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Results per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Filter by network (BTC, ETH, BSC, POL, TRX, SOL, LTC, BCH, XRP)
        in: query
        name: network
        type: string
      - description: Filter by currency (POL, ETH, USDT, USDC, …)
        in: query
        name: currency
        type: string
      - description: Filter by status (pending, confirmed, failed)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payments.GetTransactionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: WaaS on-chain transactions
      tags:
      - WaaS - Crypto Wallets
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token — prefix with "Bearer ": Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"